worker: fix require.cache + better err msg
This fixes the require.cache busting to actually work.

The previous patch fixed it for exactly two require's, but now
it does a fresh require for every start of the FakeApiGateway.

Also change the error message for rejected promise to make
it really obvouis that it's coming from this module.